+++ app/views/caracteristicas.scala.html
@@ -2,7 +2,7 @@
 @(alunoForm: Form[CaracteristicaVO])(implicit messages: Messages)
 
 @main("Novas Caracteristicas"){
- @helper.form(action = routes.HomeController.SubmeterCar()) {
+ @helper.form(action = routes.caracteristicasController.SubmeterCar()) {
 
 <table class="">
         <tbody>
